packages/bitcoincash-oauth-django/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bitcoincash-oauth-django"
version = "0.2.5"
description = "Bitcoin Cash OAuth authentication for Django and Django REST Framework"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Security :: Cryptography",
]
keywords = ["bitcoin", "bitcoin-cash", "oauth", "authentication", "django", "drf", "ecdsa"]
requires-python = ">=3.8"
dependencies = [
    "django>=4.0",
    "djangorestframework>=3.14.0",
    "coincurve>=18.0.0",
    "cashaddress>=1.0.6",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
]

[project.urls]
Homepage = "https://github.com/paytaca/bitcoincash-oauth"
Documentation = "https://github.com/paytaca/bitcoincash-oauth#readme"
Repository = "https://github.com/paytaca/bitcoincash-oauth.git"
Issues = "https://github.com/paytaca/bitcoincash-oauth/issues"

[tool.hatch.build.targets.wheel]
packages = ["bitcoincash_oauth_django"]
